mines: open surrounding on right click of numbers
When the player right clicks on an opened number, which has the
corresponding number of flags surrounding it, open the unflagged
unopened tiles around the number.  This is achieved with simultaneous
left and right click in Microsoft Minesweeper.  Note, that if the flags
are set incorrectly, the player will lose.

This is doable, because prior to this commit, right click was only
responsible for flagging and unflagging unopened tiles.