mines: remove flag from opened tiles
If a player left-clicks on a flagged tile, which doesn't have a mine, it
opens.  However, under the hood, the data attribute "flagged" remains
set to "true".  This causes bugs when player right clicks on nearby
numbers to invoke openNeighborsAroundNumber().

Ensure that open tiles are not marked with data attribute "flagged".