pairs: add card flipping animations
Rewrite pairs.js to use HTML content instead of CSS `content:`.  Relying
on CSS `content:` inside `::after` makes it impossible to use the property
`backface-visibility`, which is required for the 3D animation.

A big benefit of using this approach is that we no longer need calls to
`refreshEmoji()` in random places in the code.