pairs: fix emoji centering
Firefox on Linux seems to be fine with regard to positioning of emojis,
but in Firefox on Windows and Firefox for Android the emojis are skewed.
In Windows, the font makes emojis stick out from backside-up cards.

Add more CSS to make the emojis centered on Firefox for Android.  This
probably also fixes it on Windows (though I haven't tested it).