mines: add difficulty levels
Replace current fixed difficulty (10x10 grid with up to 10 mines) of
"mines.html" with three difficulty levels.  I couldn't come up with good
emoji representations for them, so just numbers 1, 2, 3 on the buttons
to choose the level will do for now.

Rewrite mine generation algorithm to produce a fixed number of mines
instead of "up to 10".  Simply shuffle an array of booleans, which is
filled with a number `true` and `false` values which correspond to the
number of mines and the number of empty cells.

See also: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/fill

Different difficulty levels have different grid sizes.  That means that
the <table> needs to be regenerated to game restart.  Clear the whole
HTML of the grid in constructor of class `Grid`.  "pairs.html" doesn't
care about this at the moment -- the object of class `Grid` is created
once per page load there.