mines: propagateEmpty: open tiles properly
Function propagateEmpty() doesn't update the field "flagged" in the
dataset of a tile element.  This causes a bug, steps to reproduce:

  1. place incorrect flags on tiles without mines
  2. open an empty field, which opens the tiles with the incorrect flags

After these steps right-click on the tiles adjacent to the incorrectly
flagged tiles is unusable.  Player doesn't see the flags, but they are
still being counted.  This shouldn't affect "correct gameplay", but will
definitely confuse players who are messing around.

Ensure that tiles which are opened by function propagateEmpty() delete
their flags, which are incorrect from gameplay POV.